Updated the MPC_init to create a controller using the MPCtools library
The MPC_init currently doesnt work on account of the R matrix when the
controller creates the Kalman filter with the LQR command. The error states
that the matrix, declared as "R=eye(5)" is not positive definite. When,
The R matrix that eventually arrives in the LQR funciton is in fact the zero
matrix, but I have yet to discover when R is redeclared.