Initial commit of the ROS project
The directory now includes the necessary structure for running the code,
complete with three nodes launched from the XML file in the /launch directory.
The structure contains three nodes, (1) the master node, to which commands can be
given from the terminal (should be replaced by a GUI eventually), (2) the
reference generator node and (3) the quadcopter model node. The last is used
soley to simulate the system as an alternative to runnig the actual process.
Could be used in the Kalman filter later on.

Finally, the /config directory contains a JSON configuraiton file which is
passed to the nodes in the launch file. This enables us to set attributes in
all nodes from one place, both globally and in specific nodes.