Added the reference signal described in the work of lukkonen
Some interesting nondeterministic behaviour was noted when sending the data
with ROS publishers/subscribers, and we chould consider imlpementing the
realtime_tools publishers for the (harder) control loop in future
implementations. Also, it would be useful to communicate the reference
signal from the controller with ZMQ, which will be used to send the data to
the quadcopter.