Updated configuration file, quadcopter model and dependencies
The configuration file now defines the initial states of the quadcopter
model in a 12x1 vector. The file can also be edited in the
create_configuration script, in which the encoded JSON object is stored.

The quadcopterModel node now includes the dynamics, which can be run
separately in perfect time by executing the simulation script in /examples.
However, in the current ROS implementation, the reference is passed trhough
a topic, causing non-deterministic timelags and giving rise to a slightly
messy simulation (though functional).

The dependencies in the package XML were updated to accomodate the changes.