Added kalman filter for for the linearized system (PD+quadcopter)
The EKDF filter is still dysfunctional, but the kalman filter test model runs,
albeit a bit strangely. The process noise covariance matrix in the _init file
was computed using the conv command and the measurement covariance matrix was
set to a diagonal matrix (as the noise is uncorrelated in the model) with
variances set accoring to the appled noise.