Updated report with quadcopter dynamics and included functional kalman filter
The regular discrete kalman filter applied to the linearized quadcopter model
(including PD dynamics) is funcitonal and works well. The kalman model cannot
use the ss commands, as the class doesn't support code generation. Instead,
the model was discretized using the commented commands, and the discrete A and
B matrices were simply pasted into the script. This will not be an issue in
the python implementation.

documented in the report.tex