Added module with useful classes and updated the .cnf file to include PD-parameters
The modeule quad_classes was created and contains the quadcopter class,
which takes an omega vector (at t = tk) and outputs the states and measurements
of the system (at t = tk + h) for some timestep h defined in .cnf. The
module also contains an inner PD controller and an initial commit of a kalman
filter.

The simulate_* scripts showcase how the modules are imported and used. Note
that this simplifies the ROS implementation somewhat (assuming we'll simulate
the system there at some point) and also allows us to compute the system
Jacobian more easily, so that the EKDF can be tested and compared to the
regular Kalman filter.