Initial commit of trajectory generation for changing references on the horizon
The trajectoryData.mat file is loaded in a model callback (mpc_trajectory_data_
callback.m), as simulink can't load structs of single matrices from the
workspace without including time and signal fields. A fixed step solver must
be used and the trajectory is currently lagging in phase for some reason.

The .mex solver used in the S-file is still operational, but doesn't take the
dynamic reference into account. A new solver has been generated and will be
committed as soon as the reference matrix generation works properly.