Updated crazylib with functional UKF and included UKF/EKF in the report
The crazylib module can now perform an UKF update, but the example
in /examples has yet to be completes (I will do so ASAP). Included the
quadcopter dynamics in the crazylib in order to do so and wrote about our
UKF/EKF implementations in the report to make the code somewhat comprehensible.